feat: header profile dropdown
