+++ frontend/src/components/common/HeaderIcons.tsx
@@ -1,15 +1,49 @@
+'use client';
+
 import { BsGithub } from 'react-icons/bs';
-import { MdSettings } from 'react-icons/md';
-import { useState } from 'react';
-import HeaderProfileDropdown from './HeaderProfileDropdown';
+import { MdPerson, MdSettings } from 'react-icons/md';
 import Link from 'next/link';
+import LoginButton from '../auth/LoginButton';
+import LogoutButton from '../auth/LogoutButton';
+import { useUser } from '@auth0/nextjs-auth0/client';
+
+const HeaderProfileDropdown = () => {
+  const { user } = useUser();
+
+  return (
+    <ul
+      tabIndex={0}
+      className="dropdown-content z-[1] p-2 shadow rounded-box mt-4 w-48 bg-white py-1"
+    >
+      {user ? (
+        <div>
+          <li>
+            <Link
+              href={`/profile`}
+              className="px-4 py-2 text-sm text-gray-700 flex flex-row items-center flex-start gap-2"
+            >
+              <MdPerson className="text-lg" />
+              Your Profile
+            </Link>
+          </li>
+          <li>
+            <div className="px-4 py-2 text-sm text-gray-700">
+              <LogoutButton />
+            </div>
+          </li>
+        </div>
+      ) : (
+        <li>
+          <div className="px-4 py-2 text-sm text-gray-700">
+            <LoginButton />
+          </div>
+        </li>
+      )}
+    </ul>
+  );
+};
 
 const HeaderIcons = () => {
-  const [dropDown, setDropDown] = useState(false);
-  const handleDropDown = () => {
-    setDropDown(!dropDown);
-  };
-
   return (
     <div className="absolute inset-y-0 right-0 flex items-center pr-2 sm:static sm:inset-auto sm:ml-6 ">
       <div className="relative ml-3">
@@ -20,11 +54,11 @@ const HeaderIcons = () => {
           >
             <BsGithub className="text-2xl" />
           </Link>
-          <div onClick={handleDropDown}>
-            <MdSettings className="text-2xl" />
+          <div className="dropdown dropdown-end">
+            <MdSettings tabIndex={0} className="text-2xl" />
+            <HeaderProfileDropdown />
           </div>
         </div>
-        {dropDown && <HeaderProfileDropdown />}
       </div>
     </div>
   );
